WHORYOU_Project/requirements.txt
streamlit
langchain
langchain-community
openai
tiktoken
flair
konlpy
torch
transformers
llama-cpp-python
loguru